Don't overwrite `certN.pem` certificates with Let's Encrypt certificates.
The former are provided via environment variables on the haproxy service and linked backend services. The latter are provided via the linked letsencrypt service. We should be able to mix and match them. For example, when we need to use an EV certificate for one domain but Let's Encrypt certificates are OK for other domains.