Don't explicitly set `CERT_FOLDER` to default value. See ixc/acmi#142
Doing so will cause `DEFAULT_SSL_CERT` and `SSL_CERT` environment variables
to be ignored. We want to mix and match certificates defined in environment
variables with Let's Encrypt certificates.

Instead, set `DEFAULT_SSL_CERT` to the value of the default/first Let's
Encrypt certificate (if not already explicitly set) to ensure that SSL
termination is enabled.